\name{MPCC-package}
\docType{package}
\title{
  MPCC - Matrix pearson correlation
}
\description{
  TODO
}
\details{
 TODO
}
\author{ 
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com}\cr
  Contributions from: JICS
}
\references{
  \itemize{
    \item TODO
  }
}
\seealso{
  \itemize{
    \item TODO
  }
}
\keyword{package}
