\name{PCC}
\alias{PCC}
\alias{PCC.naive}
\title{PCC - Matrix pearson correlation }
\description{
  Matrix pearson correlation
}
\usage{
PCC(aM, bM, use = NULL)
PCC.naive(aM, bM, use = NULL)
}
\arguments{
  \item{aM}{ Matrix aM. (n x m) }
  \item{bM}{ Matrix bM. (n x p) }
  \item{use}{ use = ignored, only for backwards compatibility with cor() }
}
\value{
  TODO
}
\details{
  TODO
}
\examples{
  # TODO
}
\seealso{
  TODO
}
\references{
  TODO
}
\author{ 
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\keyword{methods}
